requirements.txt
# Core framework
Django>=5.1,<5.3
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt[crypto]>=5.5,<6.0
django-cors-headers>=4.7,<5.0
django-filter>=24.3,<26.0
drf-spectacular>=0.27,<1.0
whitenoise>=6.9,<7.0

# Configuration
python-dotenv>=1.0,<2.0

# Production server
gunicorn>=23.0,<24.0

# Database (PostgreSQL — optional if using SQLite)
psycopg2-binary>=2.9,<3.0

# RAG / chat (optional — only required for the /api/v1/chat/ endpoint)
openai>=1.40,<2.0
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-openai>=0.3,<0.4
faiss-cpu>=1.8,<2.0
tiktoken>=0.7,<1.0
docx2txt>=0.9,<1.0

# Excel import
openpyxl>=3.1,<4.0

# WhatsApp client
requests>=2.32,<3.0
